Remove as c_int from byte ops, use wrapping arithmetic, fix INDEX_INDICATOR type
- Replace byte-to-c_int casts with direct u8 operations (bitmask,
  comparison, shift, bitwise OR/NOT)
- Use wrapping_add/wrapping_sub for delta encoder/decoder byte arithmetic
- Change INDEX_INDICATOR from c_int to u8 in 5 files
- Remove prob as c_int in RC probability arithmetic
- Remove unused c_int imports

+++ liblzma-rs/src/common/file_info.rs
@@ -1,5 +1,5 @@
 use crate::types::*;
-use core::ffi::{c_int, c_uint, c_void};
+use core::ffi::{c_uint, c_void};
 #[repr(C)]
 pub struct lzma_index_s {
     _opaque: [u8; 0],
@@ -154,7 +154,7 @@ unsafe extern "C" fn get_padding_size(buf: *const u8, mut buf_size: size_t) -> s
     let mut padding: size_t = 0;
     while buf_size > 0 && {
         buf_size = buf_size.wrapping_sub(1);
-        *buf.offset(buf_size as isize) as c_int == 0
+        *buf.offset(buf_size as isize) == 0
     } {
         padding = padding.wrapping_add(1);
     }
