+++ liblzma-rs/src/common/vli_encoder.rs
@@ -1,5 +1,4 @@
 use crate::types::*;
-use core::ffi::c_int;
 #[no_mangle]
 pub unsafe extern "C" fn lzma_vli_encode(
     mut vli: lzma_vli,
@@ -23,7 +22,7 @@ pub unsafe extern "C" fn lzma_vli_encode(
     vli >>= (*vli_pos).wrapping_mul(7);
     while vli >= 0x80 as lzma_vli {
         *vli_pos = (*vli_pos).wrapping_add(1);
-        *out.offset(*out_pos as isize) = (vli as u8 as c_int | 0x80) as u8;
+        *out.offset(*out_pos as isize) = vli as u8 | 0x80;
         vli >>= 7;
         *out_pos = (*out_pos).wrapping_add(1);
         if *out_pos == out_size {
